refactor(packages): Split bibtex support code and parse author earlier
Move name parsing functions borrowed from nbibtex and biblatex
compatibility rules to a "support" subfolder.
Split and parse author names earlier (at entry consolidation,
rather than each time a citation is produced).
